Removing shared mdb/stateless code. We will use private classes for the mdb container like we originally have done with the stateless container.
