+++ container/openejb-core/src/main/java/org/apache/openejb/core/mdb/MdbInstanceManager.java
@@ -17,23 +17,34 @@
 
 package org.apache.openejb.core.mdb;
 
+import org.apache.openejb.ApplicationException;
 import org.apache.openejb.BeanContext;
 import org.apache.openejb.OpenEJBException;
-import org.apache.openejb.core.instance.InstanceCreatorRunnable;
-import org.apache.openejb.core.instance.InstanceManager;
-import org.apache.openejb.core.instance.InstanceManagerData;
+import org.apache.openejb.SystemException;
+import org.apache.openejb.cdi.CdiEjbBean;
+import org.apache.openejb.core.BaseContext;
+import org.apache.openejb.core.InstanceContext;
+import org.apache.openejb.core.Operation;
+import org.apache.openejb.core.ThreadContext;
+import org.apache.openejb.core.interceptor.InterceptorData;
+import org.apache.openejb.core.interceptor.InterceptorStack;
 import org.apache.openejb.loader.Options;
 import org.apache.openejb.monitoring.LocalMBeanServer;
 import org.apache.openejb.monitoring.ManagedMBean;
 import org.apache.openejb.monitoring.ObjectNameBuilder;
 import org.apache.openejb.monitoring.StatsInterceptor;
 import org.apache.openejb.spi.SecurityService;
+import org.apache.openejb.util.DaemonThreadFactory;
 import org.apache.openejb.util.Duration;
+import org.apache.openejb.util.LogCategory;
+import org.apache.openejb.util.Logger;
 import org.apache.openejb.util.PassthroughFactory;
 import org.apache.openejb.util.Pool;
 import org.apache.xbean.recipe.ObjectRecipe;
 import org.apache.xbean.recipe.Option;
 
+import javax.ejb.ConcurrentAccessTimeoutException;
+import javax.ejb.SessionBean;
 import javax.management.Attribute;
 import javax.management.AttributeList;
 import javax.management.AttributeNotFoundException;
@@ -56,18 +67,47 @@ import javax.resource.spi.ActivationSpec;
 import javax.resource.spi.ResourceAdapter;
 import java.io.Flushable;
 import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.rmi.RemoteException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.RejectedExecutionHandler;
 import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import java.util.logging.Level;
 
+import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static javax.management.MBeanOperationInfo.ACTION;
 
-public class MdbInstanceManager extends InstanceManager {
+public class MdbInstanceManager {
+    protected static final Logger logger = Logger.getInstance(LogCategory.OPENEJB, "org.apache.openejb.util.resources");
+    protected static final Method removeSessionBeanMethod;
+
+    static { // initialize it only once
+        Method foundRemoveMethod;
+        try {
+            foundRemoveMethod = SessionBean.class.getDeclaredMethod("ejbRemove");
+        } catch (final NoSuchMethodException e) {
+            foundRemoveMethod = null;
+        }
+        removeSessionBeanMethod = foundRemoveMethod;
+    }
+
+    private final Duration accessTimeout;
+    private final Duration closeTimeout;
+    private final Pool.Builder poolBuilder;
+    private final ThreadPoolExecutor executor;
+    private final ScheduledExecutorService scheduledExecutor;
 
     private final Map<BeanContext, MdbPoolContainer.MdbActivationContext> activationContexts = new ConcurrentHashMap<>();
     private final Map<BeanContext, ObjectName> mbeanNames = new ConcurrentHashMap<>();
@@ -84,7 +124,43 @@ public class MdbInstanceManager extends InstanceManager {
                               final Duration accessTimeout, final Duration closeTimeout,
                               final Pool.Builder poolBuilder, final int callbackThreads,
                               final ScheduledExecutorService ses) {
-        super(accessTimeout, closeTimeout, poolBuilder, callbackThreads, ses);
+        this.accessTimeout = accessTimeout;
+        this.closeTimeout = closeTimeout;
+        this.poolBuilder = poolBuilder;
+        this.scheduledExecutor = ses;
+
+        if (ScheduledThreadPoolExecutor.class.isInstance(ses) && !ScheduledThreadPoolExecutor.class.cast(ses).getRemoveOnCancelPolicy()) {
+            ScheduledThreadPoolExecutor.class.cast(ses).setRemoveOnCancelPolicy(true);
+        }
+
+        if (accessTimeout.getUnit() == null) {
+            accessTimeout.setUnit(TimeUnit.MILLISECONDS);
+        }
+
+        final int qsize = callbackThreads > 1 ? callbackThreads - 1 : 1;
+        final ThreadFactory threadFactory = new DaemonThreadFactory("InstanceManagerPool.worker.");
+        this.executor = new ThreadPoolExecutor(
+                callbackThreads, callbackThreads * 2,
+                1L, TimeUnit.MINUTES, new LinkedBlockingQueue<Runnable>(qsize), threadFactory);
+
+        this.executor.setRejectedExecutionHandler(new RejectedExecutionHandler() {
+            @Override
+            public void rejectedExecution(final Runnable r, final ThreadPoolExecutor tpe) {
+
+                if (null == r || null == tpe || tpe.isShutdown() || tpe.isTerminated() || tpe.isTerminating()) {
+                    return;
+                }
+
+                try {
+                    if (!tpe.getQueue().offer(r, 20, TimeUnit.SECONDS)) {
+                        logger.warning("Executor failed to run asynchronous process: " + r);
+                    }
+                } catch (final InterruptedException e) {
+                    //Ignore
+                }
+            }
+        });
+
         this.securityService = securityService;
         this.resourceAdapter = resourceAdapter;
         this.inboundRecovery = inboundRecovery;
@@ -93,7 +169,7 @@ public class MdbInstanceManager extends InstanceManager {
 
 
     public void deploy(final BeanContext beanContext, final ActivationSpec activationSpec, final EndpointFactory endpointFactory)
-            throws OpenEJBException{
+            throws OpenEJBException {
         if (inboundRecovery != null) {
             inboundRecovery.recover(resourceAdapter, activationSpec, containerID.toString());
         }
@@ -118,7 +194,7 @@ public class MdbInstanceManager extends InstanceManager {
         final long maxAge = builder.getMaxAge().getTime(TimeUnit.MILLISECONDS);
         final double maxAgeOffset = builder.getMaxAgeOffset();
 
-        final InstanceManagerData data = new InstanceManagerData(builder.build(), accessTimeout, closeTimeout);
+        final Data data = new Data(builder.build(), accessTimeout, closeTimeout);
 
         MdbContext mdbContext = new MdbContext(securityService, new Flushable() {
             @Override
@@ -182,7 +258,7 @@ public class MdbInstanceManager extends InstanceManager {
 
             String jmxName = beanContext.getActivationProperties().get("MdbJMXControl");
             if (jmxName == null) {
-                jmxName  = "true";
+                jmxName = "true";
             }
 
             addJMxControl(beanContext, jmxName, activationContext);
@@ -209,7 +285,7 @@ public class MdbInstanceManager extends InstanceManager {
         data.getPool().start();
     }
 
-    public void undeploy(final BeanContext beanContext){
+    public void undeploy(final BeanContext beanContext) {
         final MdbPoolContainer.MdbActivationContext actContext = activationContexts.get(beanContext);
         if (actContext == null) {
             return;
@@ -335,4 +411,282 @@ public class MdbInstanceManager extends InstanceManager {
             return ATTRIBUTE_LIST;
         }
     }
+
+    private final class InstanceSupplier implements Pool.Supplier<Instance> {
+        private final BeanContext beanContext;
+
+        public InstanceSupplier(final BeanContext beanContext) {
+            this.beanContext = beanContext;
+        }
+
+        @Override
+        public void discard(final Instance instance, final Pool.Event reason) {
+
+            final ThreadContext ctx = new ThreadContext(beanContext, null);
+            final ThreadContext oldCallContext = ThreadContext.enter(ctx);
+            try {
+                freeInstance(ctx, instance);
+            } finally {
+                ThreadContext.exit(oldCallContext);
+            }
+        }
+
+        @Override
+        public Instance create() {
+            final ThreadContext ctx = new ThreadContext(beanContext, null);
+            final ThreadContext oldCallContext = ThreadContext.enter(ctx);
+            try {
+                return createInstance(ctx.getBeanContext());
+            } catch (final OpenEJBException e) {
+                logger.error("Unable to fill pool: for deployment '" + beanContext.getDeploymentID() + "'", e);
+            } finally {
+                ThreadContext.exit(oldCallContext);
+            }
+            return null;
+        }
+    }
+
+    public void destroy() {
+        if (executor != null) {
+            executor.shutdown();
+            try {
+                if (!executor.awaitTermination(10000, MILLISECONDS)) {
+                    java.util.logging.Logger.getLogger(this.getClass().getName()).log(Level.WARNING, getClass().getSimpleName() + " pool  timeout expired");
+                }
+            } catch (final InterruptedException e) {
+                Thread.currentThread().interrupt();
+            }
+        }
+        if (scheduledExecutor != null) {
+            scheduledExecutor.shutdown();
+            try {
+                if (!scheduledExecutor.awaitTermination(10000, MILLISECONDS)) {
+                    java.util.logging.Logger.getLogger(this.getClass().getName()).log(Level.WARNING, getClass().getSimpleName() + " pool  timeout expired");
+                }
+            } catch (final InterruptedException e) {
+                Thread.currentThread().interrupt();
+            }
+        }
+    }
+
+    /**
+     * Removes an instance from the pool and returns it for use
+     * by the container in business methods.
+     * <p/>
+     * If the pool is at it's limit the StrictPooling flag will
+     * cause this thread to wait.
+     * <p/>
+     * If StrictPooling is not enabled this method will create a
+     * new bean instance performing all required injection
+     * and callbacks before returning it in a method ready state.
+     *
+     * @param callContext ThreadContext
+     * @return Object
+     * @throws OpenEJBException
+     */
+    public Instance getInstance(final ThreadContext callContext) throws OpenEJBException {
+        final BeanContext beanContext = callContext.getBeanContext();
+        final Data data = (Data) beanContext.getContainerData();
+
+        Instance instance = null;
+        try {
+            final Pool<Instance>.Entry entry = data.poolPop();
+
+            if (entry != null) {
+                instance = entry.get();
+                instance.setPoolEntry(entry);
+            }
+        } catch (final TimeoutException e) {
+            final String msg = "No instances available in Session Bean pool.  Waited " + data.getAccessTimeout().toString();
+            final ConcurrentAccessTimeoutException timeoutException = new ConcurrentAccessTimeoutException(msg);
+            timeoutException.fillInStackTrace();
+            throw new ApplicationException(timeoutException);
+        } catch (final InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw new OpenEJBException("Unexpected Interruption of current thread: ", e);
+        }
+
+        if (null == instance) {
+            instance = createInstance(beanContext);
+        }
+
+        return instance;
+    }
+
+    private Instance createInstance(final BeanContext beanContext) throws ApplicationException {
+        try {
+            final InstanceContext context = beanContext.newInstance();
+            return new Instance(context.getBean(), context.getInterceptors(), context.getCreationalContext());
+
+        } catch (Throwable e) {
+            if (e instanceof InvocationTargetException) {
+                e = ((InvocationTargetException) e).getTargetException();
+            }
+            final String t = "The bean instance " + beanContext.getDeploymentID() + " threw a system exception:" + e;
+            logger.error(t, e);
+            throw new ApplicationException(new RemoteException("Cannot obtain a free instance.", e));
+        }
+    }
+
+    /**
+     * All instances are removed from the pool in getInstance(...).  They are only
+     * returned by the Container via this method under two circumstances.
+     * <p/>
+     * 1.  The business method returns normally
+     * 2.  The business method throws an application exception
+     * <p/>
+     * Instances are not returned to the pool if the business method threw a system
+     * exception.
+     *
+     * @param callContext ThreadContext
+     * @param bean        Object
+     * @throws OpenEJBException
+     */
+    public void poolInstance(final ThreadContext callContext, final Object bean) throws OpenEJBException {
+
+        if (bean == null) {
+            throw new SystemException("Invalid arguments");
+        }
+
+        final Instance instance = Instance.class.cast(bean);
+        final BeanContext beanContext = callContext.getBeanContext();
+        final Data data = (Data) beanContext.getContainerData();
+        final Pool<Instance> pool = data.getPool();
+
+        if (instance.getPoolEntry() != null) {
+            pool.push(instance.getPoolEntry());
+        } else {
+            pool.push(instance);
+        }
+    }
+
+    /**
+     * This method is called to release the semaphore in case of the business method
+     * throwing a system exception
+     *
+     * @param callContext ThreadContext
+     * @param bean        Object
+     */
+    public void discardInstance(final ThreadContext callContext, final Object bean) throws SystemException {
+
+        if (bean == null) {
+            throw new SystemException("Invalid arguments");
+        }
+
+        final Instance instance = Instance.class.cast(bean);
+        final BeanContext beanContext = callContext.getBeanContext();
+        final Data data = (Data) beanContext.getContainerData();
+
+        if (null != data) {
+            final Pool<Instance> pool = data.getPool();
+            pool.discard(instance.getPoolEntry());
+        }
+    }
+
+    @SuppressWarnings("unchecked")
+    private void freeInstance(final ThreadContext callContext, final Instance instance) {
+        try {
+            callContext.setCurrentOperation(Operation.PRE_DESTROY);
+            final BeanContext beanContext = callContext.getBeanContext();
+
+            final Method remove = instance.bean instanceof SessionBean ? removeSessionBeanMethod : null;
+
+            final List<InterceptorData> callbackInterceptors = beanContext.getCallbackInterceptors();
+            final InterceptorStack interceptorStack = new InterceptorStack(instance.bean, remove, Operation.PRE_DESTROY, callbackInterceptors, instance.interceptors);
+
+            final CdiEjbBean<Object> bean = beanContext.get(CdiEjbBean.class);
+            if (bean != null) { // TODO: see if it should be called before or after next call
+                bean.getInjectionTarget().preDestroy(instance.bean);
+            }
+            interceptorStack.invoke();
+
+            if (instance.creationalContext != null) {
+                instance.creationalContext.release();
+            }
+        } catch (final Throwable re) {
+            logger.error("The bean instance " + instance + " threw a system exception:" + re, re);
+        }
+
+    }
+
+    private void setDefault(final Duration duration, final TimeUnit unit) {
+        if (duration.getUnit() == null) {
+            duration.setUnit(unit);
+        }
+    }
+
+    private final class InstanceCreatorRunnable implements Runnable {
+
+        private final Data data;
+        private final InstanceSupplier supplier;
+        private final long offset;
+
+        public InstanceCreatorRunnable(final long maxAge, final long iteration, final long min, final double maxAgeOffset,
+                                       final Data data, final InstanceSupplier supplier) {
+            this.data = data;
+            this.supplier = supplier;
+            this.offset = maxAge > 0 ? (long) (maxAge / maxAgeOffset * min * iteration) % maxAge : 0l;
+        }
+
+        @Override
+        public void run() {
+            final Instance obj = supplier.create();
+            if (obj != null) {
+                data.getPool().add(obj, offset);
+            }
+        }
+    }
+
+    private class Data {
+
+        private final Pool<Instance> pool;
+        private final Duration accessTimeout;
+        private final Duration closeTimeout;
+        private final List<ObjectName> jmxNames = new ArrayList<ObjectName>();
+        private BaseContext baseContext;
+
+        public Data(final Pool<Instance> pool, final Duration accessTimeout, final Duration closeTimeout) {
+            this.pool = pool;
+            this.accessTimeout = accessTimeout;
+            this.closeTimeout = closeTimeout;
+        }
+
+        public Duration getAccessTimeout() {
+            return accessTimeout;
+        }
+
+        public Pool<Instance>.Entry poolPop() throws InterruptedException, TimeoutException {
+            return pool.pop(accessTimeout.getTime(), accessTimeout.getUnit());
+        }
+
+        public Pool<Instance> getPool() {
+            return pool;
+        }
+
+        public void flush() {
+            this.pool.flush();
+        }
+
+        public boolean closePool() throws InterruptedException {
+            return pool.close(closeTimeout.getTime(), closeTimeout.getUnit());
+        }
+
+        public ObjectName add(final ObjectName name) {
+            jmxNames.add(name);
+            return name;
+        }
+
+        public List<ObjectName> getJmxNames() {
+            return jmxNames;
+        }
+
+        public BaseContext getBaseContext() {
+            return baseContext;
+        }
+
+        public void setBaseContext(BaseContext baseContext) {
+            this.baseContext = baseContext;
+        }
+    }
+
 }
